+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Implementations of the traits for `OsString` and `&OsStr`
+
 ### Changed
 
 ### Removed